# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MEGA-TIK Finance & Subscribers ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscribers:
    get:
      operationId: listSubscribers
      tags: [subscribers]
      summary: List subscribers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, expired, cancelled]
        - name: area_id
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [pppoe, hotspot]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of subscribers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriberListResponse"
    post:
      operationId: createSubscriber
      tags: [subscribers]
      summary: Create subscriber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriberRequest"
      responses:
        "201":
          description: Created subscriber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
  /subscribers/{id}:
    get:
      operationId: getSubscriber
      tags: [subscribers]
      summary: Get subscriber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscriber with full details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriberDetail"
    put:
      operationId: updateSubscriber
      tags: [subscribers]
      summary: Update subscriber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriberRequest"
      responses:
        "200":
          description: Updated subscriber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
    delete:
      operationId: deleteSubscriber
      tags: [subscribers]
      summary: Delete subscriber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /subscribers/{id}/deposit:
    post:
      operationId: depositToWallet
      tags: [subscribers]
      summary: Deposit to subscriber wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: Wallet updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransaction"
  /subscribers/{id}/renew:
    post:
      operationId: renewSubscription
      tags: [subscribers]
      summary: Renew subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewalRequest"
      responses:
        "200":
          description: Renewal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Renewal"
  /subscribers/{id}/add-extra-quota:
    post:
      operationId: addExtraQuota
      tags: [subscribers]
      summary: Add extra quota to subscriber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddExtraQuotaRequest"
      responses:
        "200":
          description: Extra quota added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriberExtraQuota"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Subscriber:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        phone:
          type: string
        alternatePhone:
          type: string
        type:
          type: string
          enum: [pppoe, hotspot]
        status:
          type: string
          enum: [active, suspended, expired, cancelled]
        areaId:
          type: integer
        areaNameAr:
          type: string
        areaNameEn:
          type: string
        address:
          type: string
        walletBalance:
          type: number
        notes:
          type: string
        currentPlanId:
          type: integer
        currentPlanNameAr:
          type: string
        currentPlanNameEn:
          type: string
        expiryDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameAr
        - nameEn
        - phone
        - type
        - status
        - walletBalance
    SubscriberDetail:
      allOf:
        - $ref: "#/components/schemas/Subscriber"
        - type: object
          properties:
            services:
              type: array
              items:
                $ref: "#/components/schemas/SubscriberService"
            recentRenewals:
              type: array
              items:
                $ref: "#/components/schemas/Renewal"
            walletTransactions:
              type: array
              items:
                $ref: "#/components/schemas/WalletTransaction"
            debts:
              type: array
              items:
                $ref: "#/components/schemas/Debt"
    SubscriberService:
      type: object
      properties:
        id:
          type: integer
        subscriberId:
          type: integer
        planId:
          type: integer
        planNameAr:
          type: string
        planNameEn:
          type: string
        startDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        remainingGigabytes:
          type: number
        status:
          type: string
          enum: [active, expired, suspended]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subscriberId
        - planId
        - startDate
        - expiryDate
        - status
    CreateSubscriberRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        phone:
          type: string
        alternatePhone:
          type: string
        type:
          type: string
          enum: [pppoe, hotspot]
        areaId:
          type: integer
        address:
          type: string
        notes:
          type: string
        initialPlanId:
          type: integer
      required:
        - nameAr
        - nameEn
        - phone
        - type
    SubscriberListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subscriber"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    DepositRequest:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, vodafone_cash, bank_transfer]
        description:
          type: string
      required:
        - amount
        - paymentMethod
    RenewalRequest:
      type: object
      properties:
        planId:
          type: integer
        paymentType:
          type: string
          enum: [cash, vodafone_cash, bank_transfer, wallet, deferred]
        amount:
          type: number
        notes:
          type: string
      required:
        - planId
        - paymentType
    AddExtraQuotaRequest:
      type: object
      properties:
        extraPackageId:
          type: integer
        paymentType:
          type: string
          enum: [cash, vodafone_cash, bank_transfer, wallet, deferred]
        notes:
          type: string
      required:
        - extraPackageId
        - paymentType
    Renewal:
      type: object
      properties:
        id:
          type: integer
        subscriberId:
          type: integer
        subscriberNameAr:
          type: string
        subscriberNameEn:
          type: string
        planId:
          type: integer
        planNameAr:
          type: string
        planNameEn:
          type: string
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        paymentType:
          type: string
          enum: [cash, vodafone_cash, bank_transfer, wallet, deferred]
        renewalDate:
          type: string
          format: date-time
        newExpiryDate:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subscriberId
        - planId
        - amount
        - paymentType
        - renewalDate
        - newExpiryDate
...
    SubscriberExtraQuota:
      type: object
      properties:
        id:
          type: integer
        subscriberId:
          type: integer
        extraPackageId:
          type: integer
        packageNameAr:
          type: string
        packageNameEn:
          type: string
        gigabytes:
          type: number
        price:
          type: number
        paymentType:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subscriberId
        - extraPackageId
        - gigabytes
        - price
        - paymentType
    WalletTransaction:
      type: object
      properties:
        id:
          type: integer
        subscriberId:
          type: integer
        subscriberNameAr:
          type: string
        userId:
          type: integer
        userName:
          type: string
        type:
          type: string
          enum: [deposit, deduction, refund]
        amount:
          type: number
        balanceAfter:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subscriberId
        - type
        - amount
        - balanceAfter
...
    Debt:
      type: object
      properties:
        id:
          type: integer
        subscriberId:
          type: integer
        subscriberNameAr:
          type: string
        subscriberNameEn:
          type: string
        renewalId:
          type: integer
        amount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        status:
          type: string
          enum: [unpaid, partial, paid]
        dueDate:
          type: string
          format: date
        notes:
          type: string
        payments:
          type: array
          items:
            $ref: "#/components/schemas/DebtPayment"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subscriberId
        - amount
        - paidAmount
        - remainingAmount
        - status
...
    DebtPayment:
      type: object
      properties:
        id:
          type: integer
        debtId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, vodafone_cash, bank_transfer]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - debtId
        - amount
        - paymentMethod
...